splitted client/app.js to components
